Agent/requirements.txt
# ─── Core Framework ───
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1

# ─── LangChain / LangGraph ───
langchain==1.2.6
langchain-core==1.2.7
langchain-cohere==0.5.0
langchain-community==0.4.1
langchain-google-genai==3.2.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6

# ─── Vector Store (FAISS) ───
faiss-cpu==1.13.2

# ─── PDF Parsing ───
pypdf==6.6.0

# ─── Data Processing ───
pandas==2.3.3
numpy==2.4.1